Remove duplicates from aas-lib
