+++ data-plane-aas/build.gradle.kts
@@ -13,6 +13,8 @@ val mockserverVersion: String by project
 repositories { mavenCentral() }
 
 dependencies {
+    implementation(project(":aas-lib"))
+
     implementation("$group:data-plane-spi:${edcVersion}") {
         exclude(group = "dev.failsafe", module = "failsafe")
         exclude(group = group, module = "policy-model")
